Tests/LibWeb/Text/expected/textarea-click-caret-placement.txt

Click at start of second visual line: start=24 end=24
Click far past end of first visual line: start=23 end=23
Click far past end of second visual line: start=23 end=23
Click far past end of third visual line: start=35 end=35
Click just past end of third visual line: start=35 end=35
Click inside third visual line: start=40 end=40
Click on empty line: start=40 end=40
Click far past end of last line: start=45 end=45
